split dump and API into separate systemd services
